requirements.txt
django-channels==0.7.0 
aioredis==1.2.0
asgiref==3.1.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.5.1
Automat==0.7.0
certifi==2019.3.9
chardet==3.0.4
constantly==15.1.0
daphne==2.3.0
Django==2.2.1
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.1
oauthlib==3.0.1
Pillow==6.0.0
psycopg2==2.8.2
PyHamcrest==1.9.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.0.
txaio==18.8.1
urllib3==1.25.3
zope.interface==4.6.0
uWSGI==2.0.18
